Reorganization application to three main gropu: shop, userShop and postShop
