+++ app/templates/body.html
@@ -20,16 +20,16 @@
     <ul id="navibar">
         <li><a href="{{ url_for('shop.home') }}" id="home">Home</a></li>
         {% if current_user.is_authenticated %}
-        <li><a href="{{ url_for('shop.profile',username=current_user)}}" id="profile">Profile</a></li>
+        <li><a href="{{ url_for('userShop.profile',username=current_user)}}" id="profile">Profile</a></li>
         {% endif %}
-        <li><a href="{{ url_for('shop.posts') }}" id="posts">Posts</a></li>
+        <li><a href="{{ url_for('postShop.posts') }}" id="posts">Posts</a></li>
         <li><a href="{{ url_for('shop.shop_list') }}" id="shop_list" >List</a></li>
         <li><a href="{{ url_for('shop.cart')  }}" id="cart">Cart</a> </li>
-        <li><a href="{{ url_for('shop.users') }}" id="users">Users List</a></li>
+        <li><a href="{{ url_for('userShop.users') }}" id="users">Users List</a></li>
         {% if current_user.is_anonymous %}
-        <li><a href="{{ url_for('shop.login') }}" id="login">Login</a></li>
+        <li><a href="{{ url_for('userShop.login') }}" id="login">Login</a></li>
         {% else %}
-        <li><a href="{{ url_for('shop.logout') }}" id="logout">Logout: <b>{{ current_user}}</b></a></li>
+        <li><a href="{{ url_for('userShop.logout') }}" id="logout">Logout: <b>{{ current_user}}</b></a></li>
         {% endif %}
     </ul>
   </div>
