+++ app/templates/login.html
@@ -16,7 +16,7 @@
             <p>{{ form.submit() }}</p><br>
         </form>
             <ul>
-                <li><a href="{{ url_for('shop.register') }}" id="users">Registration</a></li>
+                <li><a href="{{ url_for('userShop.register') }}" id="users">Registration</a></li>
             </ul><br><br>
     </div>
 </body>
