+++ app/templates/profile.html
@@ -10,9 +10,9 @@
         <li>Email: <b>{{ profile.email }}</b> </li>
         <br>
     </ul>
-    <a href="{{ url_for('shop.change_password')}}" id="changepass">Change password</a>
+    <a href="{{ url_for('userShop.change_password')}}" id="changepass">Change password</a>
     <br><br><p><b>Your posts:</b></p>
-    <form  class="post-form" method = "POST" action="{{ url_for('shop.remove_post')}}">
+    <form  class="post-form" method = "POST" action="{{ url_for('postShop.remove_post')}}">
         {% for post in posts %}
         <div class="sec_posts">
             <p><b>{{post.title}}</b><input type="checkbox" name="remove_post" value={{post.id}}></p>
